kit: the lane is the first word, the hp is one type, and the nav module finally has to obey its own rule
Tranche H — the four decisions that were the owner's to make, now made.

THE LANE IS THE FIRST TOKEN (#15a). parse_replicate_info was a pairwise
constraint matrix for exactly one reason: the lane hid as an option among
tuning knobs, so every rule had to be spelled as an implication between
tokens. Three first tokens — replicate, owner, predict — give each lane a
closed option set, and three of the four cross-token rules did not get
rewritten, they became UNREPRESENTABLE: you cannot write two first tokens, so
owner-xor-predict has nothing to check, and "slack= requires predict" is gone
because slack= is simply not in the other lanes' sets. What survived are
per-lane checks about the FIELD'S TYPE, which were never the confusing part.

Proven wire-neutral, and more strongly than by fingerprint alone: the entire
generated output of eight fixtures was diffed before and after, and the only
deltas anywhere are the #load_hash CRCs of the source files whose tag strings
changed. Every Entity_Desc, entity table, command table, and NET_FINGERPRINT
byte-identical — a retagged build joins an un-retagged peer bit for bit.

entity= LEADS ITS TAG (#15b), synthesizing the export and the
resource=PackedScene it always implied. The claim was verified before being
relied on: the parser already hard-errored without that exact spelling, at
all eighteen sites. decl.odin's Slot comment had RECORDED this inconsistency
as known-and-unfixed folklore; that text stopped being true, so it was
rewritten as the stated rule — wire declarations are first tokens.

The schema earned itself back within a day: adding owner/predict rows
propagated the runtime's skip list and both unknown-tag error messages with
no hand-edited list, which is the exact surface that historically drifted.

ONE HP TYPE, AND THE REASON WRITTEN DOWN (#26b). The item said unify u16-first;
measuring said the opposite, so it went i32. play.Health had no consumers
outside a codegen fixture, the one real Predicted_Hp consumer already had zero
casts, and i32 turned out to be load-bearing: the overlay's `basis - hp_now`
and `hp_now - pending` are difference arithmetic that must go negative —
predicted damage outrunning real hp is the whole point of it. Under u16 the
first wraps to a huge positive AND PASSES ITS `> 0` GUARD, and the second makes
a dying bar read sixty-five thousand. health.odin now carries the why, not just
the type, because the next reader will ask why an always-positive quantity is
signed: everything that reads an hp works in differences, and differences are
signed.

The combat package split is refused, in the header where someone contemplating
it will read it. The fire lane's Fire.shooter is a field GAMES own — which is
why it declined the host_relay primitive — so it is a deliberately game-facing
surface, not a layering violation hiding in a header.

NAV EARNS ITS KEEP BY CHANGING (#27). cavecrawl's dweller walks a real navmesh,
and the module came out different for having a consumer. Its scariest rule —
never query inside a resimulating pass — was documentation, and documentation
does not fail builds; queries are default-denied now, asserting that an
enclosing scope claimed pass_never_resims(). An assert rather than a degrade,
because an empty path is a LEGITIMATE answer and degrading would swallow the
exact bug the module exists to prevent. cavecrawl has no sim lane today, which
is precisely the kind of reason that rots silently.

Two more findings only a real consumer produces: path_2d(node) re-walked node →
viewport → world → map every call, free for one agent and waste for a wave; and
the cursor docs implied caching that a networked game must not do, since
Dweller_Brain rides the takeover snapshot and a cached polyline would hand a
successor host routes computed against a NavigationServer it never ran.

Honest: the detour fires rarely in the acid — DWELLER_AGGRO keeps most paths
straight — and the dens were NOT moved to inflate that receipt.

Battery: 146 pass, one red — speedballnative, matching the starvation dossier
exactly (ticks advancing 3780→4380 while every view value and the resim counter
sit frozen). Green standalone at load 3.83; this tranche's generated output for
speedball is byte-identical, so it cannot be the cause.

+++ play/health.odin
@@ -17,7 +17,7 @@ package play
 // with first sight and resyncs seeding SILENTLY, so there is no birth guard and no
 // resync re-baseline ritual to write:
 //
-//   mob_health_hp_edge :: proc(g: ^Game, self: ^Mob, old, new: u16) { ... }
+//   mob_health_hp_edge :: proc(g: ^Game, self: ^Mob, old, new: i32) { ... }
 //
 // (This block used to carry its own shadow + health_step/health_sync — the kit's edge halves
 // made that whole mechanism the framework's; a block never shadows a replicated field now.)
@@ -27,26 +27,38 @@ package play
 // The fields are public; an odd policy (a revive that sets hp to a stake, a heal-to-base that
 // ignores a raised max) is an honest direct write, host-side.
 //
-// The host writes (plain replicated fields); every peer reads. u16 spans a player's
-// handful to a boss's hundreds.
+// The host writes (plain replicated fields); every peer reads.
+//
+// WHY i32 AND NOT u16, since hp is never negative: because everything that READS an hp
+// works in differences, and differences are signed. kcombat.Predicted_Hp's overlay
+// subtracts (`basis - hp_now`, `hp_now - pending`) and MUST be able to land below zero —
+// predicted damage outrunning the real hp is the whole point of it, clamped at the end.
+// The edge half above hands you `old, new` to subtract. kui.hp_refresh takes plain ints
+// on purpose. This block was the one u16 in that chain, so every pairing it advertised
+// needed a cast at the seam — and the cast is the dangerous kind: a u16 difference wraps
+// to a huge positive instead of going negative, which reads as valid. One signed type
+// end to end, and the whole chain composes with no casts and nothing to wrap. Bytes on
+// the wire are not the reason to pick the narrower type here: a game that cares ships a
+// codec (see cavecrawl's spelunker, whose i32 hp crosses as ONE byte via `wire=`).
 //
 // THE LAYERING (play is policy, kit is mechanism): the damage clamp and the heal clamp
 // delegate to kcombat.hurt/heal — the same scalar core kcombat.hit wraps for raw-field
 // games, so a block game and a bare-i32 game can never disagree on what a death is. The
 // SEEN-before-confirmed overlay (dip the displayed hp at visual contact, reconcile when
 // the wire agrees) is kit mechanism too: kcombat.Predicted_Hp — pair it with this block
-// in the game's presentation half, never inside the replicated fields.
+// in the game's presentation half, never inside the replicated fields. That pairing now
+// type-checks literally: php_note_hit/php_display speak i32, and so does h.hp.
 
 import kcombat "godot:kit/combat"
 
 Health :: struct {
-	hp:  u16 `gd:"replicate"`, // host-authoritative; 0 = dead/downed (what that MEANS is yours)
-	max: u16 `gd:"replicate"`, // stamped by health_arm; replicated so every screen can draw a bar
+	hp:  i32 `gd:"replicate"`, // host-authoritative; 0 = dead/downed (what that MEANS is yours)
+	max: i32 `gd:"replicate"`, // stamped by health_arm; replicated so every screen can draw a bar
 }
 
 // health_arm — host, at spawn / a max change (a plate part, a boss dialing its pool): stamp the
 // max and fill to it. A max change that should NOT full-heal is a direct `h.max = ...` instead.
-health_arm :: proc(h: ^Health, max: u16) {
+health_arm :: proc(h: ^Health, max: i32) {
 	h.max = max
 	h.hp = max
 }
@@ -54,7 +66,7 @@ health_arm :: proc(h: ^Health, max: u16) {
 // health_hurt — host: take `dmg`, clamped. Returns what actually landed (`dealt`, for hit credit
 // — 0 on a corpse) and whether THIS hit was the killing one (false on a corpse: a death edge
 // fires once).
-health_hurt :: proc(h: ^Health, dmg: u16) -> (dealt: u16, died: bool) {
+health_hurt :: proc(h: ^Health, dmg: i32) -> (dealt: i32, died: bool) {
 	return kcombat.hurt(&h.hp, dmg)
 }
 
@@ -65,7 +77,7 @@ health_kill :: proc(h: ^Health) {
 
 // health_heal — host: restore `amount`, clamped to max. Reviving a corpse is a heal from 0 —
 // gate "may this target be healed" yourself (that's policy).
-health_heal :: proc(h: ^Health, amount: u16) {
+health_heal :: proc(h: ^Health, amount: i32) {
 	kcombat.heal(&h.hp, h.max, amount)
 }
